% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring.R
\name{coloring}
\alias{coloring}
\alias{coloring_default}
\alias{coloring_highlight_covs}
\title{Select VA plot coloring}
\usage{
coloring_default(result, ...)

coloring_highlight_covs(result, ...)
}
\arguments{
\item{result}{A va_result object}

\item{...}{Additional arguments}
}
\value{
A vector of colors (in hexadecimal representation)
}
\description{
These functions generate color palettes for the plotting functions
}
\section{Functions}{
\itemize{
\item \code{coloring_default}: Default

\item \code{coloring_highlight_covs}: Highlights covariates
}}

