% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_va.R
\name{compute_va}
\alias{compute_va}
\title{Compute variability attribution}
\usage{
compute_va(va_input, conditioning = default_conditioning, idv = "TIME",
  facets = NULL)
}
\arguments{
\item{va_input}{An input data structure as produced by the prepare_va_* functions}

\item{conditioning}{A function or list determining in which order the computation should occur}

\item{idv}{The independent variable column}

\item{facets}{Columns to be used for facetting}
}
\value{
A data frame
}
\description{
This function uses the data structure prepared by the prepare functions and performs the actual
variability attribution computation. The function will return a data frame that can be used to
generate a plot.
}
