% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_va.R
\name{conditioning}
\alias{conditioning}
\alias{default_conditioning}
\alias{grouped_covariates_conditioning}
\title{Determine conditioning order}
\usage{
default_conditioning(va_input)

grouped_covariates_conditioning(va_input)
}
\arguments{
\item{va_input}{An input data structure as produced by the prepare_va_* functions}
}
\value{
A list of variable names
}
\description{
These functions automatically generate a conditioning order based on the available information.
}
\section{Functions}{
\itemize{
\item \code{default_conditioning}: Conditions first on each of the covariates and then on each of the random effects

\item \code{grouped_covariates_conditioning}: Conditions first jointly on all covariates and then on each of the random effects
}}

