% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_va.R
\name{default_conditioning}
\alias{default_conditioning}
\title{Determine conditioning order}
\usage{
default_conditioning(va_input)
}
\arguments{
\item{va_input}{An input data structure as produced by the prepare_va_* functions}
}
\value{
A list of variable names
}
\description{
These functions automatically generate a conditioning order based on the available information.
}
