% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_va.R
\name{plot_va}
\alias{plot_va}
\title{Plot variability attribution}
\usage{
plot_va(result, coloring = coloring_default)
}
\arguments{
\item{result}{A va_result as produced by compute_va}

\item{coloring}{Color specifications}
}
\description{
Plot variability attribution
}
