% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_va.R
\name{plot_va_compare}
\alias{plot_va_compare}
\title{Compare variability attribution plots}
\usage{
plot_va_compare(..., coloring = coloring_default, smooth = FALSE)
}
\arguments{
\item{...}{The VA result structures to be compared (names are used as a label)}

\item{coloring}{Color specifications}

\item{smooth}{Whether to smoothing should be performed or not}
}
\description{
The function allows to compare several VA plots by plotting them next to each other.
}
