% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_va_fun.R
\name{prepare_va_fun}
\alias{prepare_va_fun}
\title{Create a VA plot for an R model}
\usage{
prepare_va_fun(model_fun, theta, omega, sigma, time, z)
}
\arguments{
\item{model_fun}{The function implementing the model}

\item{theta}{Vector of theta parameters}

\item{omega}{Omega matrix}

\item{sigma}{Sigma matrix}

\item{time}{Vector of time points}

\item{z}{Vector of covariates}
}
\description{
Create a VA plot for an R model
}
