% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.R
\name{prepare_va_nm}
\alias{prepare_va_nm}
\title{Read results from a NM run}
\usage{
prepare_va_nm(lst_file, problem = NULL,
  column_specs = nm_column_specs(),
  column_mappers = nm_column_mappers())
}
\arguments{
\item{lst_file}{Path to a NONMEM results file}

\item{problem}{Number of the $PROBLEM to use}

\item{column_specs}{Column specifications as produced by nm_column_specs()}

\item{column_mappers}{Derivative to variable mapping specification as produced by nm_column_mappers()}
}
\description{
Read results from a NM run
}
