% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_input.R
\name{update_omega}
\alias{update_omega}
\title{Update Omega matrix}
\usage{
update_omega(input, omega)
}
\arguments{
\item{input}{A VA input as produced by the prepare_va_* functions}

\item{omega}{A covariance matrix}
}
\value{
An updated VA input structure
}
\description{
This function allows to update the Omega matrix associated with VA input structure.
}
