% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaplot-package.R
\docType{package}
\name{vaplot-package}
\alias{vaplot}
\alias{vaplot-package}
\title{vaplot: Variability Attribution Plots for Nonlinear Mixed Effect Models}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastianueckert/vaplot}
  \item Report bugs at \url{https://github.com/sebastianueckert/vaplot/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Ueckert \email{sebastian.ueckert@gmail.com} (0000-0002-3712-0255)

}
\keyword{internal}
