% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaplot-package.R
\name{matches}
\alias{matches}
\title{matches function}
\usage{
read_nm_derivative_results("run.lst", eta_columns = matches("^ETA"))
}
\description{
See \code{dplyr::\link[dplyr:matches]{matches}} for details.
}
\keyword{internal}
