% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_va.R
\name{plot_va}
\alias{plot_va}
\title{Plot variability attribution}
\usage{
plot_va(va_results, colors = NULL)
}
\arguments{
\item{va_results}{A va_result as produced by compute_va}

\item{colors}{Color specifications}
}
\description{
Plot variability attribution
}
