% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frem.R
\name{prepare_va_frem}
\alias{prepare_va_frem}
\title{Read results from a PsN FREM run}
\usage{
prepare_va_frem(directory, frem_specs = default_frem_specs())
}
\arguments{
\item{frem_specs}{}
}
\description{
Read results from a PsN FREM run
}
